Add sakkyoku.timer and sakkyoku.service files
This commit adds two new files, sakkyoku.timer and sakkyoku.service, which are used for scheduling and running the sakkyoku update every minute. The sakkyoku.timer file specifies the timer settings, while the sakkyoku.service file defines the service settings.

Add utils.go file to internal/utils package

This commit adds a new file, utils.go, to the internal/utils package. The utils.go file contains a utility function called CheckError, which is used to check for errors and print error messages.

Add slack.go file to internal/notifier package

This commit adds a new file, slack.go, to the internal/notifier package. The slack.go file contains the implementation of the SlackNotifier struct, which is used for sending notifications to Slack.

Add go.mod file

This commit adds a new file, go.mod, which is used for managing the project's dependencies. The go.mod file specifies the required modules and their versions.

Add config.go file to internal/config package

This commit adds a new file, config.go, to the internal/config package. The config.go file contains the implementation of the Config struct, which is used for loading and storing the project's configuration settings.

Add updater.go file to internal/updater package

This commit adds a new file, updater.go, to the internal/updater package. The updater.go file contains the implementation of the Updater struct, which is used for updating the sakkyoku managed projects.

Add main.go file to cmd/sakkyoku package

This commit adds a new file, main.go, to the cmd/sakkyoku package. The main.go file contains the entry point of the sakkyoku command-line tool and defines the available commands and their functionality.